* Added server port to be accessible via the client connection-settings.txt file. * Fixed bug concerning the parameter value type: integer is used instead of double for the SpinnerNumberModel now. Exception was thrown before. * ConnectionParameters instance is taken for the connection dialog. * Updated to MPA client version 1.0.3.

+++ src/de/mpa/client/settings/Parameter.java
@@ -239,7 +239,7 @@ public abstract class Parameter {
 			} else {
 				if (value instanceof String) {
 					try {
-						number = Double.valueOf((String) value);
+						number = Integer.valueOf((String) value);
 					} catch (Exception e) {
 						// do nothing, silently fail
 					}
